[#18] Added the possibility to deactivate persons
Created a new collection for persons to which they can be moved when they quit. This collection is used to lookup the name of the person but it will not be outputed as its own page.

This is needed since the person may have created news, blog posts or presentations that are connected to their id.

Also added a Jekyll plugin to concat arrays.